+++ src/geometry.c
@@ -50,6 +50,6 @@ void printfPoint (char* name, point a) {
         printf ("%s | x : %f ; y : %f ; z : %f\n",name,  a.x, a.y, a.z);
 }
 
-float norma (vector v) {
+double norma (vector v) {
 	return sqrt(v.x*v.x + v.y*v.y + v.z*v.z);
 }
